requirements.txt
minepy==1.2.6
numpy==1.20.1
pandas==1.2.4
PyYAML==6.0
scipy==1.8.0
soundex==1.1.3
